--- config/raw_config/forecast_config.json
@@ -1,158 +1,0 @@
-{
-    "location_id": {
-        "mapping": "city_key",
-        "data_type": "STRING",
-        "transformation": null,
-        "validation": null
-    },
-    "date": {
-        "mapping": "forecast.forecastday[].date",
-        "data_type": "DATE",
-        "transformation": null,
-        "validation": "validate_date"
-    },
-    "date_epoch": {
-        "mapping": "forecast.forecastday[].date_epoch",
-        "data_type": "INTEGER",
-        "transformation": null,
-        "validation": null
-    },
-    "maxtemp_c": {
-        "mapping": "forecast.forecastday[].day.maxtemp_c",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": "validate_temperature_celsius"
-    },
-    "maxtemp_f": {
-        "mapping": "forecast.forecastday[].day.maxtemp_f",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": "validate_temperature_fahrenheit"
-    },
-    "mintemp_c": {
-        "mapping": "forecast.forecastday[].day.mintemp_c",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": "validate_temperature_celsius"
-    },
-    "mintemp_f": {
-        "mapping": "forecast.forecastday[].day.mintemp_f",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": "validate_temperature_fahrenheit"
-    },
-    "avgtemp_c": {
-        "mapping": "forecast.forecastday[].day.avgtemp_c",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": "validate_temperature_celsius"
-    },
-    "avgtemp_f": {
-        "mapping": "forecast.forecastday[].day.avgtemp_f",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": "validate_temperature_fahrenheit"
-    },
-    "maxwind_mph": {
-        "mapping": "forecast.forecastday[].day.maxwind_mph",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": null
-    },
-    "maxwind_kph": {
-        "mapping": "forecast.forecastday[].day.maxwind_kph",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": null
-    },
-    "totalprecip_mm": {
-        "mapping": "forecast.forecastday[].day.totalprecip_mm",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": null
-    },
-    "totalprecip_in": {
-        "mapping": "forecast.forecastday[].day.totalprecip_in",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": null
-    },
-    "totalsnow_cm": {
-        "mapping": "forecast.forecastday[].day.totalsnow_cm",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": null
-    },
-    "avgvis_km": {
-        "mapping": "forecast.forecastday[].day.avgvis_km",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": null
-    },
-    "avgvis_miles": {
-        "mapping": "forecast.forecastday[].day.avgvis_miles",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": null
-    },
-    "avghumidity": {
-        "mapping": "forecast.forecastday[].day.avghumidity",
-        "data_type": "INTEGER",
-        "transformation": null,
-        "validation": "validate_humidity"
-    },
-    "daily_will_it_rain": {
-        "mapping": "forecast.forecastday[].day.daily_will_it_rain",
-        "data_type": "BOOLEAN",
-        "transformation": null,
-        "validation": "validate_boolean"
-    },
-    "daily_chance_of_rain": {
-        "mapping": "forecast.forecastday[].day.daily_chance_of_rain",
-        "data_type": "INTEGER",
-        "transformation": null,
-        "validation": null
-    },
-    "daily_will_it_snow": {
-        "mapping": "forecast.forecastday[].day.daily_will_it_snow",
-        "data_type": "BOOLEAN",
-        "transformation": null,
-        "validation": "validate_boolean"
-    },
-    "daily_chance_of_snow": {
-        "mapping": "forecast.forecastday[].day.daily_chance_of_snow",
-        "data_type": "INTEGER",
-        "transformation": null,
-        "validation": null
-    },
-    "condition_text": {
-        "mapping": "forecast.forecastday[].day.condition.text",
-        "data_type": "STRING",
-        "transformation": null,
-        "validation": null
-    },
-    "condition_code": {
-        "mapping": "forecast.forecastday[].day.condition.code",
-        "data_type": "INTEGER",
-        "transformation": null,
-        "validation": null
-    },
-    "uv": {
-        "mapping": "forecast.forecastday[].day.uv",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": null
-    },
-    "batch_id": {
-        "mapping": "batch_id",
-        "data_type": "STRING",
-        "transformation": null,
-        "validation": null
-    },
-    "load_dt": {
-        "mapping": "load_dt",
-        "data_type": "TIMESTAMP",
-        "transformation": null,
-        "validation": null
-    }
-}
