--- config/raw_config/location_config.json
@@ -1,68 +1,0 @@
-{
-    "location_id": {
-        "mapping": "city_key",
-        "data_type": "STRING",
-        "transformation": null,
-        "validation": null
-    },
-    "name": {
-        "mapping": "location.name",
-        "data_type": "STRING",
-        "transformation": null,
-        "validation": null
-    },
-    "region": {
-        "mapping": "location.region",
-        "data_type": "STRING",
-        "transformation": null,
-        "validation": null
-    },
-    "country": {
-        "mapping": "location.country",
-        "data_type": "STRING",
-        "transformation": null,
-        "validation": null
-    },
-    "latitude": {
-        "mapping": "location.lat",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": "validate_latitude"
-    },
-    "longitude": {
-        "mapping": "location.lon",
-        "data_type": "FLOAT",
-        "transformation": null,
-        "validation": "validate_longitude"
-    },
-    "timezone_id": {
-        "mapping": "location.tz_id",
-        "data_type": "STRING",
-        "transformation": null,
-        "validation": "validate_timezone"
-    },
-    "localtime_epoch": {
-        "mapping": "location.localtime_epoch",
-        "data_type": "INTEGER",
-        "transformation": null,
-        "validation": null
-    },
-    "localtime": {
-        "mapping": "location.localtime",
-        "data_type": "DATETIME",
-        "transformation": null,
-        "validation": "validation_timestamp"
-    },
-    "batch_id": {
-        "mapping": "batch_id",
-        "data_type": "STRING",
-        "transformation": null,
-        "validation": null
-    },
-    "load_dt": {
-        "mapping": "load_dt",
-        "data_type": "TIMESTAMP",
-        "transformation": null,
-        "validation": null
-    }
-}
